Create README.md
The main purpose of this report is to setup the benchmark experiment to compare 5 selected
machine learning algorithms. For each algorithm the setup and the result will be explained in
terms of cross-validation method (holdout, k-fold CV, etc.), evaluation metric (AUC, Accuracy,
etc.), hyperparameter tuning, variable selection, resampling method (over-sampling,
undersampling, etc.), etc.).
The dataset used is the same for the Kaggle competition with the aim to predict whether a
customer will subscribe to a telemarketing campaign or not.
Models Selected:
This is a classification problem. But what is a classification problem ?
A classification problem is when independent variables are continuous in nature and dependent
variable is in categorical form. Here is to categorize customers as they are going to subscribe or
not. All these problem’s answers are in categorical form i.e. Yes or No. and that is why they are
two class classification problems.
Therefore the following ML models were used:
1. Logistic Regression
2. Random Forrest
3. XGBoost
4. Support Vector Machine
5. Linear Discriminant Analysis